# minor go cleanup and some performance improvements 