Some minor improvements on sha validations 

sample benchmark code https://gist.github.com/ajayk/869fedfbcdb0a64444858c14de7c7a31

Results here 
```
goos: darwin
goarch: arm64
pkg: a.com
cpu: Apple M2 Pro
BenchmarkMatchValidShaCharsDynamic-12                  1             57416 ns/op
BenchmarkMatchValidShaCharsPrecompiled-12              1             12833 ns/op
BenchmarkMatchValidShaCharsManual-12                   1               625.0 ns/op
PASS
ok      a.com   0.233s
```